Add HtmlText component
